Simplify library
No need to have so many different files for a simple lib.
Moving to the same file makes it easier to test, package, and
consume.
